requirements.txt
livekit>=1.0
livekit-agents[openai,deepgram,cartesia,silero,turn_detector,elevenlabs]~=1.0
livekit-plugins-noise-cancellation~=0.2
python-dotenv~=1.0
twilio~=9.3
fastapi~=0.115
uvicorn[standard]~=0.34
websockets~=14.0
pandas~=2.2
openpyxl~=3.1
openai~=1.50
python-multipart~=0.0.12
